\name{BNPMIXcluster-package}
\alias{BNPMIXcluster-package}
\alias{BNPMIXcluster}
\docType{package}
\title{
\packageTitle{BNPMIXcluster}
}
\description{
\packageDescription{BNPMIXcluster}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{BNPMIXcluster}
\packageIndices{BNPMIXcluster}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{BNPMIXcluster}

Maintainer: \packageMaintainer{BNPMIXcluster}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
