% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNPMIXcluster.R
\docType{package}
\name{BNPMIXcluster}
\alias{BNPMIXcluster}
\alias{BNPMIXcluster-package}
\title{BNPMIXcluster: Model based clustering for mixed scale variables}
\description{
The \strong{BNPMIXcluster} package...
}
\section{BNPMIXcluster functions}{

\code{\link{MIXcluster}}
}

