% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_param_ex.R
\docType{data}
\name{meta_param_ex}
\alias{meta_param_ex}
\title{Metaparameters for testing the \emph{BNPMIXcluster} package}
\format{A data frame with 3 rows and 13 columns.}
\usage{
meta_param_ex
}
\description{
Values for the parameters used in the three especifications (a), (b) and (c) of the exercise in section 5.1 of the article Carmona et al. (2017).
}
\details{
A data frame with 3 rows and 13 columns. Each column is a parameter used inthe function \code{MIXclustering}.
}
\seealso{
\code{\link{MIXclustering}}, \code{\link{Y_ex_5_1}}
}
