% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIXclustering.R
\name{MIXclustering}
\alias{MIXclustering}
\title{Model based clustering for mixed scale variables.}
\usage{
MIXclustering(x, var_type, n.iter_out = 1000, n.burn = 100, n.thin = 3,
  a_fix = NULL, alpha = 0.5, d_0_a = 1, d_1_a = 1, b_fix = NULL,
  d_0_b = 1, d_1_b = 5, eta = 2, d_0_z = 2.1, d_1_z = 30, kappa = 5,
  delta = 4, d_0_mu = 2.1, d_1_mu = 30, sampling_prob = NULL,
  expansion_f = NULL, max.time = Inf)
}
\arguments{
\item{x}{Matrix or data frame containing the data to be clustered.}

\item{var_type}{Character vector that indicates the type of variable in each column of x. Three possible types:
\itemize{
  \item "\strong{c}" for continuous variables.
  \item "\strong{o}" for ordinal variables (binary and ordered categorical).
  \item "\strong{m}" for nominal variables (non-ordered categorical).
}}

\item{n.iter_out}{Number of efective iterations in the MCMC procedure for clustering.}

\item{n.burn}{Number of iterations discarded as part of the burn-in period at the beginning MCMC procedure.}

\item{n.thin}{Number of iterations discarded between two effective iterations, with the purpose of reducing the autocorrelation in the chain.}

\item{a_fix}{A numeric value to set the parameter \eqn{a} in the model. If \code{NULL} (default), the parameter \eqn{a} is assigned a prior distribution. See \code{details}.}

\item{alpha}{Hyperparameter in the prior distribution of \eqn{a}. See \code{details}.}

\item{d_0_a}{Hyperparameter in the prior distribution of \eqn{a}. See \code{details}.}

\item{d_1_a}{Hyperparameter in the prior distribution of \eqn{a}. See \code{details}.}

\item{b_fix}{A numeric value to set the parameter \eqn{b} in the model. If \code{NULL} (default), the parameter \eqn{b} is assigned a prior distribution. See \code{details}.}

\item{d_0_b}{Hyperparameter in the prior distribution of \eqn{b}. See \code{details}.}

\item{d_1_b}{Hyperparameter in the prior distribution of \eqn{b}. See \code{details}.}

\item{eta}{Tuning parameter controlling the proposal in the \emph{Metropolis-Hastings} step for \eqn{b}.}

\item{d_0_z}{Hyperparameter in the prior distribution of the variance for the latent variables. See \code{details}.}

\item{d_1_z}{Hyperparameter in the prior distribution of the variance for the latent variables. See \code{details}.}

\item{kappa}{Tuning parameter controlling the proposal in the \emph{Metropolis-Hastings} step for the variance of latent variables.}

\item{delta}{Tuning parameter controlling the proposal in the \emph{Metropolis-Hastings} step for the correlation of latent variables.}

\item{d_0_mu}{Hyperparameter in the prior distribution of the variance within each cluster. See \code{details}.}

\item{d_1_mu}{Hyperparameter in the prior distribution of the variance within each cluster. See \code{details}.}

\item{sampling_prob}{vector with the sampling probabilities \eqn{\pi_i} for each individual in case that the data come from a complex survey sample. By default \eqn{\pi_i=1}.}

\item{expansion_f}{vector with the sampling design weights or expansion factors, the reciprocal of the sampling probabilities, \eqn{w_i = 1/\pi_i}. If both \code{sampling_prob} and \code{expansion_f} are specified, preference is given to \code{sampling_prob}}

\item{max.time}{Maximum time tolerated to be spend in the sampling procedure of the Metropolis-Hastings steps. If reached the routine is stopped with \code{error}.}
}
\value{
\code{MIXclustering} returns a S3 object of class "\code{MIXcluster}".

The generic methods \code{\link{summary}} and \code{\link{plot}} are defined for this class.

An object of class "MIXcluster" is a list containing the following components:
\describe{
    \item{\code{cluster}}{vector with the cluster allocation for each each row in the data. It corresponds to the iteration which is Closest-To-Average (CTA) arrangement.}
    \item{\code{Y.cluster.summary}}{a summary of the data divided by the allocation in \code{$cluster}.}
    \item{\code{Y.var_type}}{vector with the variable types in the data.}
    \item{\code{Y.na}}{vector specifying the rows with missing values.}
    \item{\code{Y.n}}{number of rows in the data.}
    \item{\code{Y.p}}{number of variables in the data.}
    \item{\code{MC.clusters}}{matrix with the cluster allocation for each each row in the data. Each column corresponds to an effective iteration in the MCMC simulation of the model (after discarding burn-in and thinning iterations).}
    \item{\code{cluster.matrix.avg}}{average similarity matrix of size \eqn{n} by \eqn{n}.}
    \item{\code{MC.values}}{a list with the simmulated values of the chainf for the parameters \eqn{a},\eqn{b},\eqn{\Lambda},\eqn{\Omega}.}
    \item{\code{MC.accept.rate}}{a named vector with the acceptance rates for each parameter. It includes iterations that are discarded in the burn-in period and thinning.}
    \item{\code{call}}{the matched call.}
}
}
\description{
Model based clustering for mixed scale variables
}
\details{
The model consists in a bayesian non-parametric approach for clustering that is capable to combine different types of variables through the usage of associated continuous latent variables.

The clustering mechanism is based on a location mixture model with a Poisson-Dirichlet process prior on the location parameters \eqn{\mu_i ; i=1,...,n} of the asociated latent variables.

Computational inference about the cluster allocation and the posterior distribution of the parameters are performed using MCMC simulations.

The parameters \eqn{a} and \eqn{b} in the model define the \eqn{P-D} process and therefore control the number of groups. These parameters can be fixed, resulting in a larger/smaller number of groups if assigned a larger/smaller value, respectively.

There are 9 hyperparameters that characterize the prior distributions of some parameters in the model:
\itemize{
    \item \eqn{f(a)=alpha*I(a=0)+(1-alpha)*dgamma(a|d_0_a, d_1_a)}
    \item \eqn{f(b|a)=dgamma(b+a|d_0_b, d_1_b)}
    \item \eqn{sigma^2 ~ inverse-gamma(d_0_z, d_1_z)}
    \item \eqn{sigma^2_mu ~ inverse-gamma(d_0_mu, d_1_mu)}
}

The definition of these values also affect the number of resulting clusters since they affect the variance implied in the model.

For example, increasing the values of \code{d_1_a} and \code{d_1_b} reduce the number of groups.

The parameters \code{eta}, \code{kappa}, \code{delta} are tuning parameters that control the acceptance rate in the random-walk MH steps of the new proposed values for the parameters \eqn{b}, \eqn{\Lambda_jj} (variance of latents) and \eqn{\Omega_ij} (correlation of latents).

These parameters are not recomended to be changed ( functions: \code{sampling_b} , \code{sampling_Lambda_jj} , \code{sampling_Omega_ij} not exported ).
}
\examples{
# Clustering simulated data in "sim.cluster.data" #
# for more information about the data:
\dontrun{ help(sim.cluster.data) }

# exercise using 1 continuous, 1 nominal and 1 ordinal variables

Y_data <- matrix(NA,nrow=nrow(sim.cluster.data),ncol=3)
colnames(Y_data) <- paste("Y",1:3,sep=".")
Y_data[,1] <- sim.cluster.data[,2]
Y_data[,2] <- findInterval( sim.cluster.data[,3], c(-Inf,4,5,Inf) )
Y_data[,3] <- findInterval( sim.cluster.data[,4], c(-Inf,3,Inf) )

set.seed(12345) # for reproducibility

cluster_estim <- MIXclustering( Y_data,
                          var_type=c("c","m","o"),
                          n.iter_out=100, # less than the default=1000 (faster example)
                          n.burn=20, # less than the default=50
                          n.thin=3,
                          )

cluster_estim

# Summary of clustering results
summary(cluster_estim)

# Grafical representation of clustering results
plot(cluster_estim,type="heatmap")
plot(cluster_estim,type="chain",chain.obj="all")

# Comparison with the original clusters in the simulated data
plot(x=jitter(sim.cluster.data$cluster),
     y=jitter(cluster_estim$cluster),
     main="",
     xlab="Real cluster",
     ylab="Model cluster",
     pch=19, col="#FF000020")



##### Testing "MIXclustering" function with poverty.data #####
# Using entity 15 (Edomex) #

\dontrun{
Y_names <- c("ict_norm",
             "ic_ali","ic_asalud","ic_cv",
             "ic_rezedu","ic_sbv","ic_segsoc",
             "niv_ed","tam_loc")
aux_subset <- rep(T,nrow(poverty.data))
aux_subset <- aux_subset & is.element(substr(poverty.data$folioviv,1,2),"15")

poverty_cluster_estim <- MIXclustering(x=poverty.data[aux_subset,Y_names],
                                       var_type=c("c","o","o","o","o","o","o","m","m"),
                                       expansion_f = poverty.data[aux_subset,"factor_hog"],

                                       n.iter_out=1000,
                                       n.burn=100,
                                       n.thin=3
                                       )
}
}
\references{
Carmona C., Nieto-Barajas L., Canale A. (2017). \emph{Model based approach for household clustering with mixed scale variables.}
}
\seealso{
\code{\link{summary.MIXcluster}} for a summary of the clustering results, \code{\link{plot.MIXcluster}} for graphical representation of results.
}

