% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_Lambda_jj.R
\name{sampling_Lambda_jj}
\alias{sampling_Lambda_jj}
\title{MCMC sampling of parameter "\eqn{\Lambda_{j,j}}" in the \emph{mixdpcluster} model for bayesian clustering.}
\usage{
sampling_Lambda_jj(n_sim_mh = 1, sigma_jj_ini, j, d_0_z, d_1_z, kappa = 1,
  Z, mu_Z, sigma_Z, sampling_prob, max.time = 10 * 60, n.burn = 0,
  accept_display = T, verbose = F, USING_CPP = TRUE)
}
\description{
Generates a sample from the the posterior distribution of the j-th diagonal element \eqn{(i,j)} of the \eqn{\Lambda} matrix in the \emph{mixdpcluster} model for bayesian clustering.
  The simulation is done via Metropolis-Hastings method.
}
\keyword{internal}
