% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.MIXcluster.R
\name{summary.MIXcluster}
\alias{summary.MIXcluster}
\title{Summarizing clustering results}
\usage{
\method{summary}{MIXcluster}(object, ...)
}
\arguments{
\item{object}{an object of class "\code{MIXcluster}"}

\item{...}{further arguments passed to or from other methods.}
}
\description{
\code{summary} method for class "\code{MIXcluster}".
}
\seealso{
\code{\link{MIXclustering}}
}
